requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
colorama==0.4.1
cryptography==2.6.1
Django==2.1.7
djangorestframework==3.9.2
emoji==0.5.1
future==0.17.1
idna==2.8
isort==4.3.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
pycparser==2.19
pylint==2.3.1
python-telegram-bot==11.1.0
pytz==2018.9
requests==2.21.0
six==1.12.0
typed-ast==1.3.1
urllib3==1.24.1
wrapt==1.11.1
